Braces and Whitespace
Java generally puts the open brace on the same line the class, method, condition or loop declaration that opens it. The closing brace is placed on its own line, at teh same indentation level as the declaration.

A space separates the arguments to operators (+, -, *, /, %…) and pretty much any ',' including method parameters and direct declaration of arrays.