Names
They say two things are hard in software engineering: caching, and naming things.

Deciding what to name things is hard, but once you've decided it, writing it down follows some basic rules. Most identifiers in Java use camel case, that is the first letter of each word making up the name begins with a capital letter. The first letter of the identifier is a capital in the case of class names, and lower case for methods, instance variables, and local variables.

The one big exception is the case of constants and enum values, which are in all caps to indicate their unchanging nature. You separate multiple words with an underscore, as you can see in JFrame.EXIT_ON_CLOSE.