Enums/Magic Ints
You want to avoid plain integers in your code because they lack meaning. We call these magic ints, and they make your code harder to read and understand, since their meaning isn't always obvious. Fortunately there are a couple ways to use integer values and mark their meaning for you and anyone who is reading your code.

The first is to declare a constant. Each language has a way of doing this, but In java this is done via the static and final modifiers, ie "static final int". The static makes it a class level declaration, and the final means that it cannot be modified. By convention, the identifier should be in all caps.

The second way is via an enum. Enums are a way to express the idea of a discrete list of items/choices/types. In this case, you had a search type in your data. Rather than expressing these as ints we can leverage the type system in java and ensure that any parameter for that particular type is a valid value. It also lets the programmer more easily know what kind of value that parameter is. Enums are full on classes in Java, and so you can attach methods and other values to them, though this is limited in a couple important ways we will see later.