Code Reuse
You always want to avoid copying and pasting code in different places, so that if you need to change the way the code works, you don't ahve to change it multiple places. In this case, you had three listeners doing the exact same thing, only on different fields. In this case, it makes sense to put that action in its own class, and then pass the different fields into parameters in the constructor.