Anonymous Classes
Sometimes we only need an object to encapsulate a discrete action, especially in the case of a listener on a button. As a convenience, Java gives us anonymous class, which allows us to define a class inline without naming it. You can see here we've written new ActionListener, which shouldn't be possible. However, Java takes that and makes us a class automatically that implements the ActionListener interface. Since the clear text action is unlikely to be used elsewhere, we can pass in an instance of the anonymous class, allowing us to easily see what the listener on that button does.

 It is useful to think of these classes as passing in an action (a block of code) rather than passing in an object since in other languages you will be passing around actual functions rather than anonymous classes.